Changes to the Admin/Creator-only edit checks.
* Create a new check around whether it's only the user (or an admin) who
can make changes to a user-centric object. This occurs in GenericObject.
* Make it so the Object_*_Demo classes don't need to explicitly state
that they are able to write, when it's not an admin user. This change
also occurs in GenericObject.
* As a demo, the Tag class has associated config values relating to
whether only admins can create a tag, and whether the creator of the tag
is the only person who can edit that tag.