A huge bundle of changes to try to bring compliance with the Zend CodeSniffer tests, and to fix some of the Mess Detector issues.
* Lots (but not all) of the protected variables now have _ prefixes.
* Many of the lines - especially comments - now fit under 80 characters but
some may not where this would prevent the code from executing or being
readable. More of the lines fit under 120 characters, but this is still not yet
complete.
* Most non-camel-caps variable names now are in camelCaps
* The isCreationAction variable in the construct function throws an exception
when this arrives at the GenericObject class. This is because it shouldn't
be going all the way back up the stack to here - it implies insufficient
planning and thought process. Plus, it should be tracked by the Unit Tests :)
* OnlyXMayModify values are now _reqXToMod to ensure long variable names aren't
being used throughout the code base.
* Replaced some "and"'s with "&&"'s and "or"'s with "||"'s to try to improve
code speed. Also, looks nicer.

I've probably missed some stuff, but at least I'm trying to track this with
28 files changes, 1042 insertions and 875 deletions. That's a LOT.