Various changes, fixing some array logic, getting to passed Unit Tests and enabling sorting talks.
* Created basic tests for Object_Room
* Created basic tests for talk sorting
* Fixed array data in Object_AttendeeTest
* Added room capacity to Object_Room, and a function to broker by room sizes
* Added object caching to Generic_Object::broker functions which don't currently
support it.
* Moved application code from the LimboTalks plugin to the Talk class.
* Lots of changes to the talk class, including, adding intAllocatedRoomID value,
checking whether the number of attendees exceeds the capcity of the room in the
getSelf() function, added a function to build an empty grid, marking off the
rooms and slots which are locked, added a function to sort and place talks by
number of attendees, added a function to unschedule talks without a minimum
number of attendees.