Fixed several issues to get to a point where at least *one* of the edit screens is working.
* Added a new Object_Talk.html.tpl file, and it's component part -
Object_Talk.tpl.
* Made GenericObject return all the items from the source table in the "data"
value in pointers. Also, return the source value in lower case (which means
we can point the user back to that object, if necessary).
* Removed irrelevant Resources type from Talk object. If the user wants those
resources, they should speak to an organiser... for now! Also, edited the
translation strings on both the Talk object and the User object.
* Updated index.php to correct typos, and to prevent it from trying to commit
empty values in JSON fields.