This change implements features which should close #132.
This implements a new Object type - "ChangeLog", which stores the
diff(ish) of each change, the actor who performed said change, and on
which table this occurred.

This Object type is instantiated by the Generic_Object function
writeChangeLog() that bundles up the required fields, and creates the
object. This object is then overloaded by the SecureConfig and UserAuth
objects to prevent leakage of private data. This is also overloaded by
the ChangeLog object to prevent it from writing its own data over and
over again!

To call the Generic_Object function, there is a new Plugin called,
again, ChangeLog which triggers on the Create, Update and Delete
actions, after the entry has been written.